Remove editable-tree-1 from API
